OS X compatible
`readlink -f` is a GNU extension not available everywhere. After a bit
digging into this snippet (from stackoverflow), this is a really nice
trick. The problem is that it spawns quite a few subshells, but… hey,
this script will compile GHC anyway, so gaining 0.1s is NOT worth the
effort :)